+++ CHANGELOG.md
@@ -16,10 +16,16 @@ Version 0.5.5 *(2015-09-15)* Stickerpipe
 Version 0.6.3 *(2015-09-24)* Stickerpipe
 ----------------------------
 
-* Added GCM module support 
+* Added GCM module support
 * Fixes and improvements
 
 Version 0.1.3 *(2015-09-24)* GCM integration
 ----------------------------
 
-* Base GCM functionality 
+* Base GCM functionality
+
+Version 0.7.3 *(2015-10-23)* Stickerpipe
+----------------------------
+
+* Removed builders and color setters form fragments and activities
+* Customization by colors and drawable resources